refactor!: move VHI modules into myogestic/vhi/ (drop myogestic.interfaces)
