+++ tools/gen_proto.py
@@ -2,12 +2,12 @@
 """Regenerate the VHI gRPC Python stubs from the vendored .proto.
 
 The canonical contract lives in the Virtual-Hand-Interface repo
-(``proto/myogestic_vhi.proto``); ``myogestic/_proto/myogestic_vhi.proto`` is a
+(``proto/myogestic_vhi.proto``); ``myogestic/vhi/_proto/myogestic_vhi.proto`` is a
 vendored copy. After updating the vendored copy, run:
 
     uv run --extra grpc python tools/gen_proto.py
 
-This writes ``myogestic/_proto/myogestic_vhi_pb2.py`` and ``..._pb2_grpc.py``,
+This writes ``myogestic/vhi/_proto/myogestic_vhi_pb2.py`` and ``..._pb2_grpc.py``,
 which are committed so a plain install needs only grpcio at runtime (not
 grpcio-tools).
 """
@@ -19,7 +19,7 @@ import sys
 from pathlib import Path
 
 REPO_ROOT = Path(__file__).resolve().parent.parent
-PROTO_DIR = REPO_ROOT / "myogestic" / "_proto"
+PROTO_DIR = REPO_ROOT / "myogestic" / "vhi" / "_proto"
 PROTO_FILE = PROTO_DIR / "myogestic_vhi.proto"
 
 
@@ -45,7 +45,7 @@ def main() -> int:
 
     # grpc_tools emits a flat `import myogestic_vhi_pb2` in the _pb2_grpc file;
     # rewrite it to a package-relative import so the stubs work as
-    # `myogestic._proto.*`.
+    # `myogestic.vhi._proto.*`.
     grpc_file = PROTO_DIR / "myogestic_vhi_pb2_grpc.py"
     text = grpc_file.read_text()
     patched = text.replace(
